requirements.txt
# requirements.txt - DEPENDÊNCIAS DO AGENTE QA
# LOCAL: / (Raiz do projeto no Railway)

Flask==3.0.0
requests==2.31.0
mysql-connector-python==8.2.0
openai==1.12.0
gunicorn==21.2.0